reorg spatial distance & speed
